+++ packages/munchkins/src/registry/index.ts
@@ -1,1 +1,3 @@
+export { discoverAgents } from "./discover.js";
+export { registerListLaunchableCommand } from "./list-launchable-command.js";
 export { AgentRegistry, type RegisteredCommand, registry } from "./registry.js";